WIP Porting: Further refactoring
Moved Pulse and Item to root level
