+++ ecpy_pulses/sequences/views/template_view.enaml
@@ -19,7 +19,7 @@ from ecpy.utils.widgets.qt_completers import QtLineCompleter
 from ecpy.utils.widgets.dict_editor import (DictEditor, FieldFieldEditor,
                                             FieldFieldCompleterEditor)
 
-from ...workspace.views.sequence_views import (VarEditor,
+from .sequence_views import (VarEditor,
                                                instantiate_context_view)
 
 enamldef TemplateSequenceView(GroupBox): view:
